Refactor code (it'll be easirer to complete feature-3 when big main view divided by small units): Create seperate controls for every menu tab
