+++ src/Domain/DiagramApp.Domain/Toolbox/ToolboxCategory.cs
@@ -5,5 +5,6 @@
         Shapes,
         Blocks,
         Other,
+        Advanced,
     }
 }
