+++ src/Presentation/DiagramApp.Client/Components/AdvancedToolboxView.xaml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Frame xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+       xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+       xmlns:toolbox="clr-namespace:DiagramApp.Domain.Toolbox;assembly=DiagramApp.Domain"
+       xmlns:viewmodel="clr-namespace:DiagramApp.Client.ViewModels"
+       x:Class="DiagramApp.Client.Components.AdvancedToolboxView"
+       x:DataType="viewmodel:MainViewModel">
+    <Frame.Resources>
+        <ResourceDictionary>
+            <!--Refactor DRY l8r-->
+            <DataTemplate x:Key="toolboxItems"
+                          x:DataType="toolbox:ToolboxItem">
+                <Grid Padding="10"
+                      RowDefinitions="Auto, Auto">
+                    <Path Grid.Row="0"
+                          Data="{Binding PathData, Converter={StaticResource PathDataToGeometryConverter}}"
+                          Style="{StaticResource Shape}"
+                          ToolTipProperties.Text="{Binding Name}">
+                    </Path>
+                </Grid>
+            </DataTemplate>
+        </ResourceDictionary>
+    </Frame.Resources>
+    <Grid RowDefinitions="Auto, *"
+          RowSpacing="10">
+        <Label Grid.Row="0"
+               Text="Продвинутый инструментарий"
+               HorizontalTextAlignment="Center"/>
+        <CollectionView Grid.Row="1"
+                        SelectionMode="Single"
+                        ItemsSource="{Binding ToolboxViewModel.AdvancedToolboxItems}"
+                        ItemTemplate="{StaticResource toolboxItems}">
+            <CollectionView.ItemsLayout>
+                <GridItemsLayout Orientation="Vertical" Span="3"/>
+            </CollectionView.ItemsLayout>
+        </CollectionView>
+    </Grid>
+</Frame>
