+++ src/Presentation/DiagramApp.Client/Components/AdvancedToolboxView.xaml.cs
@@ -1,0 +1,9 @@
+namespace DiagramApp.Client.Components;
+
+public partial class AdvancedToolboxView : Frame
+{
+	public AdvancedToolboxView()
+	{
+		InitializeComponent();
+	}
+}
