+++ src/Presentation/DiagramApp.Client/Components/ControlsView.xaml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Frame xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:models="clr-namespace:DiagramApp.Domain.Canvas;assembly=DiagramApp.Domain"
+             xmlns:viewmodel="clr-namespace:DiagramApp.Client.ViewModels"
+             x:Class="DiagramApp.Client.Components.ControlsView"
+             x:DataType="viewmodel:MainViewModel"
+             Padding="0">
+    <StackLayout Orientation="Horizontal"
+                 Spacing="5">
+        <Grid RowDefinitions="*, Auto"
+              ColumnDefinitions="*"
+              RowSpacing="10"
+              Margin="10">
+            <HorizontalStackLayout Spacing="5">
+                <VerticalStackLayout Spacing="5"
+                                     Padding="0">
+                    <ImageButton Source="{AppThemeBinding Light=select.png, Dark=select_dark.png}"
+                                 Command="{Binding ChangeControlsCommand}"
+                                 CommandParameter="Select"
+                                 Aspect="AspectFit"
+                                 Style="{StaticResource Tool}">
+                        <ImageButton.Triggers>
+                            <DataTrigger TargetType="ImageButton"
+                                     Binding="{Binding CurrentCanvas.Controls}"
+                                     Value="{x:Static models:ControlsType.Select}">
+                                <Setter Property="Background" Value="{AppThemeBinding Light={StaticResource Gray200}, Dark={StaticResource Gray600}}"/>
+                            </DataTrigger>
+                        </ImageButton.Triggers>
+                    </ImageButton>
+                    <Label Text="Выбрать"
+                           Style="{StaticResource LabelControlName}"
+                           HorizontalOptions="Center"/>
+                </VerticalStackLayout>
+                <VerticalStackLayout Spacing="5"
+                                     Padding="0">
+                    <ImageButton Source="{AppThemeBinding Light=drag.png, Dark=drag_dark.png}"
+                                 Command="{Binding ChangeControlsCommand}"
+                                 CommandParameter="Drag"
+                                 Aspect="AspectFit"
+                                 Style="{StaticResource Tool}">
+                        <ImageButton.Triggers>
+                            <DataTrigger TargetType="ImageButton"
+                                         Binding="{Binding CurrentCanvas.Controls}"
+                                         Value="{x:Static models:ControlsType.Drag}">
+                                <Setter Property="Background" Value="{AppThemeBinding Light={StaticResource Gray200}, Dark={StaticResource Gray600}}"/>
+                            </DataTrigger>
+                        </ImageButton.Triggers>
+                    </ImageButton>
+                    <Label Text="Двигать"
+                           Style="{StaticResource LabelControlName}"
+                           HorizontalOptions="Center"/>
+                </VerticalStackLayout>
+            </HorizontalStackLayout>
+            <Label Text="Управление"
+                   Style="{StaticResource LabelControl}"
+                   Grid.Row="1" 
+                   Grid.ColumnSpan="2"
+                   VerticalOptions="Center" 
+                   HorizontalOptions="Center" />
+        </Grid>
+        <BoxView Style="{StaticResource VerticalSeparator}"/>
+    </StackLayout>
+</Frame>
