+++ src/Presentation/DiagramApp.Client/Components/ControlsView.xaml.cs
@@ -1,0 +1,9 @@
+namespace DiagramApp.Client.Components;
+
+public partial class ControlsView : Frame
+{
+	public ControlsView()
+	{
+		InitializeComponent();
+	}
+}
