+++ src/Presentation/DiagramApp.Client/Components/EditorTabsView.xaml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<CollectionView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+                xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                xmlns:viewmodel="clr-namespace:DiagramApp.Client.ViewModels"
+                xmlns:wrappers="clr-namespace:DiagramApp.Client.ViewModels.Wrappers"
+                x:Class="DiagramApp.Client.Components.EditorTabsView"        
+                x:DataType="viewmodel:MainViewModel"
+                ItemsLayout="HorizontalList"
+                ItemsSource="{Binding Canvases}">
+    <CollectionView.ItemTemplate>
+        <DataTemplate x:DataType="wrappers:ObservableCanvas">
+            <Button Text="{Binding Settings.FileName}"
+                    Command="{Binding Source={RelativeSource AncestorType={x:Type viewmodel:MainViewModel}}, Path=SelectCanvasCommand}"
+                    CommandParameter="{Binding .}"
+                    Style="{StaticResource Tab}">
+                <FlyoutBase.ContextFlyout>
+                    <MenuFlyout>
+                        <MenuFlyoutItem Text="Закрыть"
+                                        Command="{Binding Source={RelativeSource AncestorType={x:Type viewmodel:MainViewModel}}, Path=CloseCanvasCommand}"
+                                        CommandParameter="{Binding .}"/>
+                    </MenuFlyout>
+                </FlyoutBase.ContextFlyout>
+                <Button.Triggers>
+                    <DataTrigger TargetType="Button"
+                                 Binding="{Binding IsSelected}"
+                                 Value="True">
+                        <Setter Property="Opacity" Value="0.5"/>
+                    </DataTrigger>
+                </Button.Triggers>
+            </Button>
+        </DataTemplate>
+    </CollectionView.ItemTemplate>
+</CollectionView>
