+++ src/Presentation/DiagramApp.Client/Components/EditorTabsView.xaml.cs
@@ -1,0 +1,9 @@
+namespace DiagramApp.Client.Components;
+
+public partial class EditorTabsView : CollectionView
+{
+	public EditorTabsView()
+	{
+		InitializeComponent();
+	}
+}
