+++ src/Presentation/DiagramApp.Client/Components/ExplorerView.xaml.cs
@@ -1,0 +1,9 @@
+namespace DiagramApp.Client.Components;
+
+public partial class ExplorerView : Frame
+{
+	public ExplorerView()
+	{
+		InitializeComponent();
+	}
+}
