+++ src/Presentation/DiagramApp.Client/Components/PropertiesView.xaml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Frame xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+             xmlns:viewmodel="clr-namespace:DiagramApp.Client.ViewModels"
+             x:Class="DiagramApp.Client.Components.PropertiesView"
+             x:DataType="viewmodel:MainViewModel">
+    <!--Add validation l8r-->
+    <Grid RowDefinitions="Auto, *"
+          RowSpacing="10">
+        <Label Grid.Row="0"
+               HorizontalTextAlignment="Center"
+               Text="Свойства"/>
+        <Grid Grid.Row="1"
+              ColumnDefinitions="Auto, *"
+              RowDefinitions="Auto, Auto, Auto"
+              IsVisible="{Binding CurrentCanvas.SelectedFigure, Converter={toolkit:IsNotNullConverter}, FallbackValue={x:Boolean False}}"
+              ColumnSpacing="10"
+              RowSpacing="5">
+
+            <Label Text="Название: " Grid.Row="0" Grid.Column="0"/>
+            <Label Text="Изменить размер на: " Grid.Row="1" Grid.Column="0"/>
+            <Label Text="Поворот: " Grid.Row="2" Grid.Column="0"/>
+
+            <Entry Text="{Binding CurrentCanvas.SelectedFigure.Name, Mode=TwoWay}" Grid.Row="0" Grid.Column="1"/>
+            <Entry Text="{Binding CurrentCanvas.SelectedFigure.Size, Mode=TwoWay}" Grid.Row="1" Grid.Column="1"/>
+            <Entry Text="{Binding CurrentCanvas.SelectedFigure.Rotation, Mode=TwoWay}" Grid.Row="2" Grid.Column="1"/>
+        </Grid>
+    </Grid>
+</Frame>
