+++ src/Presentation/DiagramApp.Client/Components/PropertiesView.xaml.cs
@@ -1,0 +1,9 @@
+namespace DiagramApp.Client.Components;
+
+public partial class PropertiesView : Frame
+{
+	public PropertiesView()
+	{
+		InitializeComponent();
+	}
+}
