+++ src/Presentation/DiagramApp.Client/Components/ToolboxView.xaml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Frame xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:toolbox="clr-namespace:DiagramApp.Domain.Toolbox;assembly=DiagramApp.Domain"
+             xmlns:viewmodel="clr-namespace:DiagramApp.Client.ViewModels"
+             x:Class="DiagramApp.Client.Components.ToolboxView"
+             x:DataType="viewmodel:MainViewModel">
+    <Frame.Resources>
+        <ResourceDictionary>
+            <!--Refactor DRY l8r-->
+            <Style x:Key="themeAwarePathStyle" TargetType="Path" BasedOn="{StaticResource Shape}">
+                <Style.Triggers>
+                    <DataTrigger TargetType="Path"
+                     Binding="{Binding Source={x:Reference ThemeCheckBox}, Path=IsChecked}"
+                     Value="true">
+                        <Setter Property="Stroke" Value="{AppThemeBinding Dark={StaticResource OffBlack}, Light={StaticResource White}}"/>
+                        <Setter Property="Fill" Value="{AppThemeBinding Dark={StaticResource White}, Light={StaticResource OffBlack}}"/>
+                    </DataTrigger>
+                </Style.Triggers>
+            </Style>
+            <DataTemplate x:Key="toolboxItems"
+                          x:DataType="toolbox:ToolboxItem">
+                <Grid Padding="10"
+                      RowDefinitions="Auto, Auto">
+                    <Path Grid.Row="0"
+                          Data="{Binding PathData, Converter={StaticResource PathDataToGeometryConverter}}"
+                          Style="{StaticResource themeAwarePathStyle}"
+                          ToolTipProperties.Text="{Binding Name}">
+                    </Path>
+                </Grid>
+            </DataTemplate>
+        </ResourceDictionary>
+    </Frame.Resources>
+    <Grid RowDefinitions="Auto, Auto, *"
+          ColumnDefinitions="*, Auto"
+          RowSpacing="10">
+        <Label Grid.Row="0"
+               Grid.ColumnSpan="2"
+               Text="Инструментарий"
+               HorizontalTextAlignment="Center"/>
+        <Picker Grid.Row="1"
+                Grid.Column="0"
+                ItemsSource="{Binding ToolboxViewModel.Categories}"
+                SelectedItem="{Binding ToolboxViewModel.SelectedCategory}">
+            <Picker.Behaviors>
+                <toolkit:EventToCommandBehavior EventName="Loaded"
+                                                Command="{Binding ToolboxViewModel.LoadToolboxCommand}"/>
+                <toolkit:EventToCommandBehavior EventName="SelectedIndexChanged"
+                                                Command="{Binding ToolboxViewModel.CategoryChangeCommand}"/>
+            </Picker.Behaviors>
+        </Picker>
+        <CheckBox x:Name="ThemeCheckBox"
+                  Grid.Row="1"
+                  Grid.Column="1"
+                  ToolTipProperties.Text="Изменить тему объектов на альтернативную"/>
+        <CollectionView Grid.Row="2"
+                        Grid.ColumnSpan="2"
+                        SelectionMode="Single"
+                        SelectionChanged="OnSelectionChanged"
+                        ItemsSource="{Binding ToolboxViewModel.FilteredToolboxItems}"
+                        ItemTemplate="{StaticResource toolboxItems}">
+            <CollectionView.ItemsLayout>
+                <GridItemsLayout Orientation="Vertical" Span="3"/>
+            </CollectionView.ItemsLayout>
+        </CollectionView>
+    </Grid>
+</Frame>
