+++ src/Presentation/DiagramApp.Client/Resources/Styles/Styles.xaml
@@ -2,7 +2,8 @@
 <?xaml-comp compile="true" ?>
 <ResourceDictionary 
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml">
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:components="clr-namespace:DiagramApp.Client.Components">
 
     <Style TargetType="Path" x:Key="Shape">
         <Setter Property="Stroke" Value="{AppThemeBinding Dark={StaticResource White}, Light={StaticResource OffBlack}}"/>
@@ -189,12 +190,21 @@
         <Setter Property="WidthRequest" Value="250"/>
     </Style>
 
-    <Style TargetType="Frame">
+    <Style TargetType="Frame" x:Key="DefaultFrame">
         <Setter Property="BackgroundColor" Value="{AppThemeBinding Light={StaticResource Gray100}, Dark={StaticResource Gray900}}"/>
         <Setter Property="HasShadow" Value="False" />
         <Setter Property="BorderColor" Value="{AppThemeBinding Light={StaticResource Primary}, Dark={StaticResource PrimaryDark}}" />
         <Setter Property="CornerRadius" Value="0" />
     </Style>
+
+    <Style TargetType="Frame" BasedOn="{StaticResource DefaultFrame}"/>
+    
+    <Style TargetType="components:EditorView" BasedOn="{StaticResource DefaultFrame}"/>
+    <Style TargetType="components:ControlsView" BasedOn="{StaticResource DefaultFrame}"/>
+    <Style TargetType="components:ToolboxView" BasedOn="{StaticResource DefaultFrame}"/>
+    <Style TargetType="components:AdvancedToolboxView" BasedOn="{StaticResource DefaultFrame}"/>
+    <Style TargetType="components:ExplorerView" BasedOn="{StaticResource DefaultFrame}"/>
+    <Style TargetType="components:PropertiesView" BasedOn="{StaticResource DefaultFrame}"/>
 
     <Style TargetType="ImageButton">
         <Setter Property="BackgroundColor" Value="Transparent"/>
