+++ src/Presentation/DiagramApp.Client/ViewModels/ToolboxViewModel.cs
@@ -7,13 +7,19 @@ namespace DiagramApp.Client.ViewModels
 {
     public partial class ToolboxViewModel : ObservableObject
     {
+        public ToolboxCategory[] Categories => Enum.GetValues<ToolboxCategory>()
+            .Where(item => item != ToolboxCategory.Advanced)
+            .ToArray();
+        
         private readonly IToolboxService _toolboxService;
 
         public List<ToolboxItem>? ToolboxItems { get; private set; }
-        public ToolboxCategory[] Categories => Enum.GetValues<ToolboxCategory>();
 
         [ObservableProperty]
         private List<ToolboxItem>? _filteredToolboxItems;
+
+        [ObservableProperty]
+        private List<ToolboxItem>? _advancedToolboxItems;
 
         [ObservableProperty]
         private ToolboxCategory _selectedCategory;
@@ -24,14 +30,22 @@ namespace DiagramApp.Client.ViewModels
         private async Task LoadToolboxAsync()
         {
             ToolboxItems = await _toolboxService.GetToolboxItemsAsync("toolboxData.json");
+            InitializeAdvancedToolbox();
             CategoryChange();
         }
 
         [RelayCommand]
         private void CategoryChange()
         {
-            FilteredToolboxItems = ToolboxItems! // maybe contain all filteredtoolboxitems in array or smth (optimize idea) so there will be no need in filtering it again
+            FilteredToolboxItems = ToolboxItems! // maybe contain all filteredtoolboxitems in array or smth (optimize idea) so there will be no need in filtering it over and over again
                 .Where(item => item.Category == SelectedCategory)
+                .ToList();
+        }
+
+        private void InitializeAdvancedToolbox()
+        {
+            AdvancedToolboxItems = ToolboxItems!
+                .Where(item => item.Category == ToolboxCategory.Advanced)
                 .ToList();
         }
     }
